functions/requirements.txt
Flask==3.1.2
firebase-functions==0.1.2
firebase_admin==7.1.0
flask-cors
google-cloud-firestore